Readme.md
This report presents an analysis of retail sales data with the objective of gaining insights and making 
predictions regarding customer behavior and sales trends. The dataset used for this analysis consists of sales 
records from a supermarket. The report explores various aspects of the data, including data cleaning, 
exploratory data analysis, and classification tasks to predict customer type. Different models such as Logistic 
Regression, Decision Tree, Random Forest, and Support Vector Machines (SVM) are employed for the 
classification tasks. The accuracy of each model is evaluated, and a comparison is made to determine the bestperforming model.